Refactoring MoviePage component and subornitated components
