+++ src/common/utils/baseURL.js
@@ -1,0 +1,1 @@
+export const baseURL = "https://api.themoviedb.org/3";
