+++ src/common/utils/imageURL.js
@@ -1,0 +1,1 @@
+export const imageURL = "https://image.tmdb.org/t/p/original";
