+++ src/features/moviePage/Movie/styled.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const PageWrapper = styled.div`
+  max-width: 100%;
+  display: flex;
+  flex-flow: column;
+  align-items: center;
+`;
