--- src/features/moviePage/MoviePage/styled.js
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const Test = styled.p`
-  font-size: 25px;
-  font-weight: bold;
-`;
