+++ src/features/moviePage/MoviePageApp.js
@@ -1,7 +1,7 @@
 import { MoviePage } from "./MoviePage";
 
-function MoviePageApp() {
-  return (<MoviePage />);
+function MoviePageApp(id = "212") {
+  return (<MoviePage id={id} />);
 };
 
 export default MoviePageApp;
