--- src/features/moviePage/Page/styled.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const PageWrapper = styled.div`
-  max-width: 100%;
-  display: flex;
-  flex-flow: column;
-  align-items: center;
-`;
