+++ src/features/moviePage/useMovieData.js
@@ -1,0 +1,17 @@
+import { imageURL } from "../../common/utils/imageURL";
+
+export const useMovieData = ({ movieBase, movieImages, moviePeople }) => {
+
+  const posterBig = `${imageURL}/${movieImages.backdrops[1].file_path}`;
+  const posterSmall = `${imageURL}/${movieImages.posters[16].file_path}`;
+  const description = movieBase.overview;
+  const title = movieBase.title;
+  const year = "2025";
+  const production = "Poland, UK, France";
+  const date = movieBase.release_date;
+  const cast = moviePeople.cast.slice(0, 12);
+  const crew = moviePeople.crew.slice(0, 12);
+
+  return { posterBig, posterSmall, title, year, description, production, date, cast, crew };
+
+};
